% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor.mtest}
\alias{cor.mtest}
\title{Compute the matrix of correlation p-values}
\usage{
cor.mtest(x, ...)
}
\arguments{
\item{x}{numeric matrix or data frame}

\item{...}{other arguments to be passed to the function cor.test.}
}
\description{
Compute the matrix of correlation p-values
}
