% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorrplot.mixed.R
\name{ggcorrplot.mixed}
\alias{ggcorrplot.mixed}
\title{Visualize a correlation matrix using mixed methods}
\usage{
ggcorrplot.mixed(
  corr,
  upper = c("circle", "square", "ellipse", "number"),
  lower = c("number", "square", "ellipse", "circle"),
  p.mat = NULL,
  sig.lvl = 0.05,
  number.digits = 2,
  insig = c("pch", "blank"),
  pch = 4,
  pch.cex = 5
)
}
\arguments{
\item{corr}{a correlation matrix to be visualized}

\item{upper}{a character indicating the visualization method of the upper triangular matrix to be used. Currently, it supports four methods, named "circle" (default), "square", "ellipse", "number".}

\item{lower}{a character indicating the visualization method of the lower triangular matrix to be used. Currently, it supports four methods, named "circle", "square", "ellipse", "number"(default).}

\item{p.mat}{a matrix of p-value}

\item{sig.lvl}{significant level, which is 0.05 by default.}

\item{number.digits}{the number of decimal digits while the visualization method is "number".}

\item{insig}{a character specialized insignificant correlation coefficients, "pch" (default), or "blank".}

\item{pch}{a point character indicating the shape of insignificant correlation coefficients.}

\item{pch.cex}{a number controlling the shape size of insignificant correlation coefficients.}
}
\description{
Visualize a correlation matrix using mixed methods
}
