% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorrplot.R
\name{plot.method}
\alias{plot.method}
\title{Add the geom of correlation matrix to the existing plot}
\usage{
\method{plot}{method}(p, data, method = c("circle", "square", "ellipse", "number"))
}
\arguments{
\item{p}{a ggplot object}

\item{data}{a data.frame to be plotted}

\item{method}{a character indicating the visualization method of correlation matrix to be used. Currently, it supports four methods, named "circle" (default), "square", "ellipse", "number".}
}
\description{
Add the geom of correlation matrix to the existing plot
}
