% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorrplot.mixed.R
\name{ggcorrplot.mixed}
\alias{ggcorrplot.mixed}
\title{Visualize a correlation matrix using mixed methods}
\usage{
ggcorrplot.mixed(
  corr,
  upper = c("circle", "square", "ellipse", "number"),
  lower = c("number", "square", "ellipse", "circle"),
  col = NULL,
  p.mat = NULL,
  sig.lvl = 0.05,
  number.digits = 2,
  insig = c("pch", "blank", "label_sig"),
  pch = 4,
  pch.cex = 5
)
}
\arguments{
\item{corr}{a correlation matrix to be visualized}

\item{upper}{a character indicating the visualization method of the upper triangular matrix to be used. Currently, it supports four methods, named \code{"circle"} (default), \code{"square"}, \code{"ellipse"}, \code{"number"}.}

\item{lower}{a character indicating the visualization method of the lower triangular matrix to be used. Currently, it supports four methods, named \code{"circle"}, \code{"square"}, \code{"ellipse"}, \code{"number"}(default).}

\item{col}{a vector of the colors to be used, which are distributed uniformly from -1 to 1. If NULL, col will be set to \code{RColorBrewer::brewer.pal(n = 11, name = "RdBu")}, the default colour scheme of `corrplot`.}

\item{p.mat}{a matrix of p-value}

\item{sig.lvl}{a numeric vector specifying significant level(s). If the p-value in \code{p.mat} is bigger than \code{sig.lvl} (0.05 by default), then the corresponding correlation coefficient is regarded as insignificant. If \code{insig} is \code{"label_sig"}, this may be an increasing vector of significance levels, for example \code{c(0.05, 0.01, 0.001)}, in which case \code{pch} will be used once for the highest p-value interval and multiple times (e.g. "*", "**", "***") for each lower p-value interval.}

\item{number.digits}{the number of decimal digits (2 by default) while the visualization method is \code{"number"}.}

\item{insig}{a character specialized insignificant correlation coefficients, \code{"pch"} (default), \code{"blank"}, or \code{"label_sig"}.}

\item{pch}{a point character indicating the shape of insignificant correlation coefficients.}

\item{pch.cex}{a number controlling the shape size of insignificant correlation coefficients.}
}
\description{
Visualize a correlation matrix using mixed methods
}
