% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorrplot.R
\name{ggcorrplot}
\alias{ggcorrplot}
\title{Visualize a correlation matrix}
\usage{
ggcorrplot(
  corr,
  method = c("circle", "square", "ellipse", "number"),
  type = c("full", "lower", "upper"),
  p.mat = NULL,
  sig.lvl = 0.05,
  number.digits = 2,
  show.diag = TRUE,
  insig = c("pch", "blank"),
  pch = 4,
  pch.cex = 5
)
}
\arguments{
\item{corr}{a correlation matrix to be visualized}

\item{method}{a character indicating the visualization method of correlation matrix to be used. Currently, it supports four methods, named "circle" (default), "square", "ellipse", "number".}

\item{type}{a character indicating that "full" (default) matrix, "upper" or "lower" triangular matrix is displayed.}

\item{p.mat}{a matrix of p-value}

\item{sig.lvl}{significant level, which is 0.05 by default.}

\item{number.digits}{the number of decimal digits while the visualization method is "number".}

\item{show.diag}{a logical indicating whether display the correlation coefficients on the principal diagonal.}

\item{insig}{a character specialized insignificant correlation coefficients, "pch" (default), or "blank".}

\item{pch}{a point character indicating the shape of insignificant correlation coefficients.}

\item{pch.cex}{a number controlling the shape size of insignificant correlation coefficients.}
}
\description{
Visualize a correlation matrix
}
